Update dolib.rb
Added the numerical sorting algorithm as a monkey patch for the Array class.  Added a few hiera instances to look up the primary domain name for each server in a cluster as "domain" and the dhcp server domain from hiera as "dhcpserver"

Added further granularity to auditing all the systems in the cluster to be piped through the aha command line tool to be emailed.